requirements.txt
Pillow
numpy
keras